requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
cryptography==41.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
